requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
nltk>=3.8.1
joblib>=1.3.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.13.0
Flask>=3.0.0
